Added an array containing details of the attached files to the page information.
We add an empty list for all previous versions of the page and then
populate the list for the final one.  The idea is that the processing
code will copy each attachment over and add it to git as we process each page.

Note the use of urllib.quote in order to discover the filesystem name
which trac gives to the filename